requirements.txt
diffusers==0.21.0
pytorch-lightning==2.1.4
numpy==1.22.4
controlnet-aux==0.0.6
huggingface-hub==0.20.2
transformers==4.36.2
tqdm==4.65.2
accelerate==0.26.1
einops==0.7.0
tensorboardx==2.6
trimesh==3.12.7
scikit-image==0.22.0
pyopengl==3.1.4
scipy==1.11.4
torch-ema==0.3
imageio==2.33.1
imageio-ffmpeg==0.4.9
pyyaml==6.0
setuptools==60.2.0
matplotlib==3.4.2
omegaconf==2.3.0
pymeshlab==2022.2.post3
rich==13.4.2
ninja==1.11.1.1
git+https://github.com/NVlabs/nvdiffrast.git
git+https://github.com/openai/CLIP.git